pyproject.toml
[tool.poetry]
name = "lab1"
version = "0.1.0"
description = ""
authors = ["Natalia Lopez <natalia.lopezg@udea.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.4"
matplotlib = "^3.8.3"
pillow = "^10.2.0"
scikit-learn = "^1.4.1.post1"
pandas = "^2.2.0"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
